[feature] AccountScreen: redesigned tariff section to show tariff history and support dual rates
