+++ composeApp/src/commonMain/composeResources/values/strings.xml
@@ -85,6 +85,7 @@
     <string name="tariffs_direct_debit_monthly">Direct Debit Monthly</string>
     <string name="tariffs_error_load_tariffs">Unable to retrieve tariffs</string>
     <string name="tariffs_tariff_effective">Rates effective %1$s</string>
+    <string name="tariffs_variable_rate">Variable Unit Rate</string>
 
     <!-- Retail Regions -->
     <string name="retail_region_eastern_england">Eastern England</string>
@@ -131,9 +132,11 @@
     <string name="account_mpan">MPAN: %1$s</string>
     <string name="account_error_null_tariff">Unable to retrieve your tariff details.</string>
     <string name="account_tariff_start_date">Tariff starts on %1$s</string>
-    <string name="account_tariff_end_date">Tariff ends on %1$s</string>
-    <string name="account_tariff_unit_rate">Standard Unit Rate\n(p/kWh)</string>
-    <string name="account_tariff_standing_charge">Standing Charge\n(p/day)</string>
+    <string name="account_tariff_date_range">Tariff from %1$s to %2$s</string>
+    <string name="account_tariff_unit_rate">Standard Unit Rate (p/kWh)</string>
+    <string name="account_tariff_day_unit_rate">Day Unit Rate (p/kWh)</string>
+    <string name="account_tariff_night_unit_rate">Night Unit Rate (p/kWh)</string>
+    <string name="account_tariff_standing_charge">Standing Charge (p/day)</string>
     <string name="account_meter_serial">Meter: %1$s</string>
     <string name="account_meter_selected">SELECTED</string>
     <string name="account_meter_select">SELECT</string>
